feat: Implement LLM Rubric Scoring API with comprehensive testing
