+++ api/app.py
@@ -195,5 +195,5 @@ def get_rubric_info():
     }), 200
 
 if __name__ == '__main__':
-    port = int(os.environ.get('PORT', 5000))
+    port = int(os.environ.get('PORT', 5001))  # Changed from 5000 to 5001
     app.run(host='0.0.0.0', port=port, debug=True)
